+++ CHANGELOG.rst
@@ -12,6 +12,22 @@ The semantic versioning only considers the public API as described in
 paths are considered internals and can change in minor and patch releases.
 
 
+v4.35.0 (2024-12-??)
+--------------------
+
+Changed
+^^^^^^^
+- Argument groups created from dataclass-like that have zero configurable
+  arguments no longer adds a config loader (`#???
+  <https://github.com/omni-us/jsonargparse/pull/???>`__).
+
+Deprecated
+^^^^^^^^^^
+- ``add_dataclass_arguments`` is deprecated and will be removed in v5.0.0.
+  Instead use ``add_class_arguments`` (`#???
+  <https://github.com/omni-us/jsonargparse/pull/???>`__).
+
+
 v4.34.1 (2024-11-??)
 --------------------